### 2017-06-14 #126 @maelle
rtimicropem
This package aims at supporting the analysis of PM2.5 measures made with RTI MicroPEM. [RTI MicroPEM](https://www.rti.org/sites/default/files/brochures/rti_micropem.pdf) are personal monitoring devices (PM2.5 and PM10) developped by [RTI international](https://www.rti.org/).

```
Package: rtimicropem
Title: Supports the Analysis of RTI MicroPEM Output Files
Version: 1.3
Authors@R: c(person("Maëlle", "Salmon", email = "maelle.salmon@yahoo.se", role = c("aut", "cre")),
             person("Zheng", "Zhou", comment = "Department of Environmental Health Sciences Mailman School of Public Health at Columbia University, New York, USA", role = c("aut")),
             person(family = "ERC Grant Agreement number 336167 - the CHAI Project", role = c("fnd")),
             person("Carles", "Milà", role = c("ctb")),
             person("Sreekanth", "Vakacherla", role = c("ctb")),
             person("Cathryn", "Tonne", role = c("ctb")),
             person("Julian", "Marshall", role = c("ctb")))
Description: Supports the input and reproducible analysis of RTI MicroPEM output files.
Suggests: testthat,
    xtable,
    shiny
Imports: lubridate,
   rbokeh,
    R6,
   changepoint,
    lazyeval,
    pathological,
    methods

```

https://github.com/masalmon/rtimicropem

[e.g., "data extraction, because the package parses a scientific data file format"]

This package fits in the data extraction category because it allows to transform files produced by the RTI MicroPEM, which are scientific devices, into data.frames.

Scientists using RTI MicroPEMs for collecting personal PM2.5 data.

Not to my knowledge.
