### 2017-06-16 #126 @sckott
### Editor checks:

---

#### Editor comments

Thanks for your submission @maelle  !

Currently seeking reviewers. It's a good fit and not overlapping.

* `goodpractice::gp()` output is below. Looks quite clean.

```r
It is good practice to

    R/clean_measures.R:56:NA
    R/clean_measures.R:74:NA
    R/clean_measures.R:75:NA
    R/clean_measures.R:76:NA
    R/clean_measures.R:77:NA
    ... and 26 more lines

    inst/shiny-examples/myapp/server.R:50:1
    inst/shiny-examples/myapp/ui.R:42:1
    R/identify_lags.R:20:1
    R/identify_lags.R:25:1
    R/micropem_class.R:236:1

  ✖ not import packages as a whole, as this can cause name
    clashes between the imported packages. Instead, import only the
    specific functions you need.
  ✖ fix this R CMD check NOTE: Namespace in Imports field not
    imported from: ‘R6’ All declared Imports should be used.
```

---

Reviewers: xx xx
Due date: xxxxx
