### 2017-06-21 #126 @maelle
@sckott The package no longer imports `ggplot2` as a whole but I don't understand where the R CMD check NOTE comes from.

`fix this R CMD check NOTE: Namespace in Imports field not
    imported from: ‘R6’ All declared Imports should be used.`

In micropem-class.R I actually import `R6::R6Class` :thinking:
