### 2017-06-21 #126 @sckott
Right - i get that error sometimes too and doesn’t seem to make sense - will look into that warning message in general cause I'd like to know
