### 2017-06-22 #126 @maelle
* I apparently don't get that note on Travis, and I didn't get it on my PC, probably linked to the R CMD check options that `goodpractice` applies? The only NOTE I get is `checking DESCRIPTION meta-information ... NOTE
Authors@R field gives persons with no valid roles:
  ERC Grant Agreement number 336167 - the CHAI Project [fnd]` which was expected, and the output of `goodpractice::gp()` is

```r
── GP rtimicropem ──────────────────────────────────────────────────────────────────────────────

It is good practice to

  ✖ write unit tests for all functions, and all package code in general. 92% of code
    lines are covered by test cases.

    R/clean_measures.R:56:NA
    R/clean_measures.R:74:NA
    R/clean_measures.R:75:NA
    R/clean_measures.R:76:NA
    R/clean_measures.R:77:NA
    ... and 26 more lines

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer editor
    windows that are about 80 characters wide. Try make your lines shorter than 80
    characters

    inst\shiny-examples\myapp\server.R:50:1
    inst\shiny-examples\myapp\ui.R:42:1
    R\identify_lags.R:20:1
    R\identify_lags.R:25:1
    R\micropem_class.R:235:1

────────────────────────────────────────────────────────────────────────────────────────────────
```

And it seems I'm using `goodpractice` latest version, weird.

* The build on Mac R-devel on Travis [fails](https://travis-ci.org/maelle/rtimicropem/jobs/245307208) but I wasn't too worried about it, it looked like one of these dependencies issues that get fixed over time... Maybe I should actually worry? :thinking:
