### 2017-06-22 #126 @sckott
talked to some people about the warning about Imports and they said to just import the package, either all of it, or importFrom

hmm, little experience with osx builds, not sure
