### 2017-07-02 #126 @LucyMcGowan
>##### Paper (for packages co-submitting to JOSS)
>
>The package contains a `paper.md` with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [x] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

🙌 🎉 💯

This is an excellent package for streamlining the analysis of PM2.5 measures made with RTI MicroPEM. It seems this would be very useful for labs utilizing these methods. I have a few small comments below, but overall big 👍!

* The tests ran successfully on OS X Sierra 3.3.3
* All code in the README, vignette, and examples ran as expected.
* I get a small note with `devtools::check()`, but perhaps that is because my R is outdated? It seems `fnd` should be allowed according to utils docs.
 > Authors@R field gives persons with no valid roles:  ERC Grant Agreement number 336167 - the CHAI Project [fnd]

#### Code/API

* The function naming convention general seems consistent, except `batch_convert()` which seems to follow the `object_verb()` style, but then there is also `convert_output()` which is the opposite (although that seems to fit better with the rest of the package’s naming convention of `verb_object()`, such as `clean_measures()` and `identify_lags()`).

*These are mostly style suggestions, feel free to take them or leave them*

* I suggest adding links to functions that are referenced. For example in the `clean_measures()` details, the change point `cpt.mean()` function is referenced - it would be awesome if it had a hyperlink to view that function.
* Indicate the parameter type in the documentation (for example `Character` for a `path` parameter). Functions I noticed without input type specified:
    * `convert_output()`
    * `batch_convert()`
    * `identify_lags()`
* I wonder if the function returns nothing (such as `run_shiny_app()` and `batch_convert()` if the roxygen `@return` should just not be used rather than stating it returns nothing in the documentation? My inclination is that often if nothing is returned, the **Value** heading just isn’t used, but I could be mistaken.
* For `batch_convert()` perhaps instead of “converting the settings…to csv” it should specify “writing the settings…to csv” or “saving the settings…to csv” or “exporting the settings…to csv”? For some reason the word “convert” seemed confusing to me.
* Also for `batch_convert()` I’m not sure I understand what is meant by “See the data in inst/data to see which one applies.” but perhaps this is just because I am not familiar with MicroPEM output?
* Describe the built in shiny app a bit more in `run_shiny_app()` — maybe stating something about how this is an app for field work?

#### Other suggestions
* The README doesn’t have a title with the package name (I think this is usually typical?)
* There is a file `find_zeros.R` that I believe is blank.
