### 2017-07-04 #126 @karawoo
>##### Paper (for packages co-submitting to JOSS)
>
>The package contains a `paper.md` with:
>
>- [X] **A short summary** describing the high-level functionality of the software
>- [X] **Authors:**  A list of authors with their affiliations
>- [X] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [X] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

This seems like a very useful package for working with data from RTI MicroPEM devices. After reviewing the package I wish I had data from one of these devices so that I could visualize it with this package!  Also the function `chai_alarm()` makes me smile :tea: :alarm_clock:

All of my comments are pretty minor.

- The `\dontrun{}` section in `micropem_class.R` is missing the closing curly brace.

- Could you include a link or description of CHAI when it's first referenced in `vignette_ammon`? There is a link down in the Shiny app section, but I think it would be helpful if it came earlier since this vignette is the main introduction to the package.
- There is a typo in `chai_data_cleaning_vignette` - in the section "Removal of entire files": "there were to few" should be "there were too few".
- Also a small typo in `vignette_ammon`: "loosing" -> "losing".
- I think there's a mistake in the image that shows the MicroPEM output file in `vignette_ammon`. It says the green section goes into a field called `"control"`, but shouldn't it be `"settings"`?
- Some of the images in `vignette_ammon`, particularly the Shiny screenshots, are very big and are cut off on my laptop screen. Would it be possible to resize those?

#### Style

- Internal functions aren't marked with `#' @noRd` as the rOpenSci [style guide](https://github.com/ropensci/onboarding/blob/328434b5e37898e1e321e40f1f13e4bbe78f3d09/packaging_guide.md#-documentation) suggests, but I don't think that's a big deal.

- All tests pass on my machine.
- Overall coverage as reported by `covr::package_coverage()`: 85.41%
- Unlike @LucyMcGowan I don't get any notes when running `devtools::check()`. I'm on R 3.4.0.

#### Miscellaneous

- I also noticed the empty `find_zeros.R` file.
