### 2017-07-13 #126 @maelle
Thanks again @LucyMcGowan @karawoo! This was really helpful and I'm really glad you reviewed the package! I've now made all changes (except `#' @noRd` for internal functions, I hope it's ok @sckott ).

Lucy and Kara, are you ok with my adding you as "rev" in `DESCRIPTION` to acknowledge your contribution?
