### 2017-07-17 #126 @maelle
Added Lucy and Kara as "rev" + added the review badge.
