### 2017-07-17 #126 @sckott
Approved!

To-dos:
- [ ] Transfer the repo to rOpenSci, you know the drill :)
- [x] You've already added reviewers in DESCRIPTION as "rev"

- [ ] Add the "Peer-reviewed" badge to your README:

- [ ] Fix any links in badges for CI and coverage to point to the ropensci URL
- [ ] Submit to JOSS using the Zenodo DOI.  We will tag it for expedited review.

Welcome aboard once again! Do you want to do a blog post about? Short or long if so?
