### 2017-07-18 #126 @maelle
@sckott thanks! I transferred the repo, updated CI links etc, did the Zenodo stuff and submitted to JOSS.

I'd love to do a blog post about `rtimicropem` with some general message reg. scientific devices with weird output format and no central documentation. :-)
